Cipher Translator
No one asked but believe it or not this font (Lingojam_cipher) makes the SGA writing at lingojam more legible. There are some issues, but it does work.

You can use the template to make a font to translate it into Latin or some other script if you like.

I will add it to the main font (SGA-Regular) at some future date.

https://lingojam.com/StandardGalacticAlphabet

 #enchantingtable